disable recursion in rescan
LONG "DT/LONG"

RAMX,AMX,LONG,... results in DT/DT/DT/DT/LONG recursion. This is a
common pattern for flags. So we need to block recursion.

Because the expansion currently does a whole line at a time, we need
keep track of any macros used on the whole line and then color *all*
of them "blue". We can't color them blue at use, otherwise a
subsequent use in the same line, which is not recursive would be
blocked as well.

This is all a kind of a hack, but does work for DT/LONG and similar.