implement hint file reading and writing
The purpose is constrain address assignment to match existing code.

- add option to generate .HNT/.hnt files
- add option to parse .HNT/.hnt files

To generate a skeleton hint file, run with --create-hints.
That file can then be edited to update the hints. Hints in the file can be
of 6 types:
  - A: ucode assigned
  - C: ucode constrained
  - H: ucode hinted (sequentially allocated)
  - B: begin constraint
  - I: inner contraint
  - E: end contraint

Of these, only H and B are used/required when reading the file back in.