update constraint processing to match micro2 results
When searching for a base address to start a contraint block, do not
require all 2**n addresses to be free. Instead, simulate the action of the
inner constraints and early termination and only require that the addresses
*actually used* are free.

Otherwise it would not be possible to allocate certain blocks to the
addresses that the original microcode actually used.